+++ weblate/trans/backups.py
@@ -574,6 +574,15 @@ class ProjectBackup:
             # Create components
             self.load_components(zipfile, self.restore_component)
 
+        # Fixup linked components
+        old_slug = f"/{self.data['project']['slug']}/"
+        new_slug = f"/{project.slug}/"
+        for component in self.project.component_set.filter(
+            repo__istartswith="weblate:"
+        ):
+            component.repo = component.repo.replace(old_slug, new_slug)
+            component.save()
+
         return self.project
 
     def store_for_import(self):